requirements.txt
discord.py[voice]
asyncio
cogs
python-dotenv
requests
pillow
colorthief
discord-py-slash-command
